refactor: Consolidate BigQuery worker and remove monitoring service
- Remove separate BigQuery monitor application
- Integrate monitoring functionality into BigQuery worker
- Simplify project structure and package scripts
- Update Terraform configuration for Cloud Run service
- Remove redundant monitoring-related files and scripts
