--- apps/bigquery-monitor/package.json
@@ -1,15 +1,0 @@
-{
-    "name": "bigquery-monitor",
-    "version": "0.0.1",
-    "scripts": {
-        "dev": "bun --watch run src/index.ts",
-        "start": "bun run src/index.ts"
-    },
-    "dependencies": {
-        "@google-cloud/bigquery": "^7.0.0",
-        "@google-cloud/pubsub": "^4.0.0"
-    },
-    "devDependencies": {
-        "typescript": "^5.0.0"
-    }
-}
