--- apps/bigquery-monitor/terraform/variables.tf
@@ -1,11 +1,0 @@
-variable "project_id" {
-  description = "GCP Project ID"
-  type        = string
-  default     = "dev-iot-452706"
-}
-
-variable "region" {
-  description = "GCP Region"
-  type        = string
-  default     = "asia-northeast1"
-}
