+++ apps/bigquery-worker/src/models/sensors.ts
