+++ apps/iot-client/src/simulator.ts
@@ -1,6 +1,12 @@
 
 import { PubSub } from '@google-cloud/pubsub';
-import { fetchDevices } from './devices';
+
+const mock = {
+    // 模拟的传感器数量
+    deviceCount: 2,
+    // 每个设备的传感器数量
+    sensorsPerDevice: 2,
+}
 
 // 验证环境变量
 const credentialsPath = process.env.GOOGLE_APPLICATION_CREDENTIALS;
@@ -66,9 +72,20 @@ async function main() {
     // 验证认证
     await verifyAuthentication();
 
+    const deviceSensors: Array<{ deviceId: string, sensorId: string }> = [];
+
+    for (let deviceNum = 1; deviceNum <= mock.deviceCount; deviceNum++) {
+        const deviceId = `device-${deviceNum.toString().padStart(3, '0')}`;
+
+        for (let sensorNum = 1; sensorNum <= mock.sensorsPerDevice; sensorNum++) {
+            const sensorId = `${deviceId}-sensor-${sensorNum.toString().padStart(2, '0')}`;
+            deviceSensors.push({ deviceId, sensorId });
+        }
+    }
+
     // 为所有设备创建定时发送任务
     setInterval(() => {
-        fetchDevices().forEach(({ deviceId, sensorId }) => {
+        deviceSensors.forEach(({ deviceId, sensorId }) => {
             const data = generateSensorData(deviceId, sensorId);
             publishMessage(data);
         });
